[x] nepenthes 0.1.5 commit
unpacked package to trunk, svn added new files and dirs.

+++ modules/submit-xmlrpc/XMLRPCParser.cpp
@@ -345,7 +345,7 @@ void dumpNodeTree(Node *tree, uint32_t depth)
 		for( i = 0; i < depth; i++ )
 			printf("  ");
 		
-		printf("%s  %08x:", n->m_key, (uint32_t)n);
+		printf("%s  %08x:", n->m_key, (unsigned int)((intptr_t)n));
 		
 		if( n->m_type == n_value )
 		{
